replace geometries module/functions with features module/functions
